pyproject.toml
[tool.poetry]
name = "graia"
version = "0.1.0"
description = ""
authors = ["GeekHaozai <1700118540@qq.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
graia-ariadne = {extras = ["standard"], version = "^0.11.5"}
requests = "^2.31.0"
bs4 = "^0.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
